Improve tuple standard conformance
Some improvements to tuple's standard conformance:
- Value constructor now takes const references as input;
- get<>() now handles all types of references correctly.

In the process of implementing this, some other functionalities were
added:
- remove_reference(_t)<>;
- move() and forward().

Unit tests for all functionalities were also added. To facilitate this,
an options was implemented to use a different namespace than std.  This
was necessary to be able to use the host's STL in the test code.